major refactor, kinda working
