+++ configs/rl/reverse_text.toml
@@ -4,13 +4,13 @@ model = "willcb/Qwen2.5-0.5B-Reverse-SFT"
 id = "reverse-text"
 
 [inference]
-gpus = 7
+gpus = 6
 
 [inference.args]
 enforce_eager = true
 
 [trainer]
-gpus = 1
+gpus = 2
 
 [trainer.args]
 run_name = "reverse-text"
